% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsLogit.R
\name{brms.binary}
\alias{brms.binary}
\title{Stan binary Model}
\usage{
brms.binary(
  data = dataActive,
  IV = IV,
  DV = DV,
  chains = 4,
  iter = 2000,
  warmup = 1000,
  cores = 10,
  seed = 1234,
  ...
)
}
\arguments{
\item{Specify}{DV, IV, data, and other arguments native to BRMS. DV must be binary, consisting of two levels.}
}
\value{
brms model object, binary regression (Logit)
}
\description{
Stan binary Model
}
