% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsLinear.R
\name{brms.linear}
\alias{brms.linear}
\title{Stan Linear Model}
\usage{
brms.linear(
  data = dataActive,
  IV = IV,
  DV = DV,
  chains = 4,
  iter = 2000,
  warmup = 1000,
  cores = 10,
  seed = 1234,
  ...
)
}
\arguments{
\item{Specify}{DV, IV, data, and other arguments native to BRMS}
}
\value{
brms model object
}
\description{
Stan Linear Model
}
