% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPrediction.R
\name{ggPoint}
\alias{ggPoint}
\title{Generate plot of point estimates from BRMS object}
\usage{
ggPoint(
  data = plotData,
  size_x = 10,
  size_y = 10,
  size_title = 12,
  title = "Authoritarianism \\nand Party Identification",
  xtitle = "Authoritarianism",
  caption = "Data come from the 2020 Western Survey",
  ytitle = "Probability",
  xlimits = c(0, 1),
  ylimits = c(0, 1),
  facet_order = "default",
  ...
)
}
\arguments{
\item{data}{{dataframe used in brms model}}

\item{model}{{brms model}}

\item{xvar}{{independent variable}}

\item{xrange}{{values of xvar}}

\item{mvar}{{moderator variable}}

\item{mrange}{{values of mvar}}

\item{category_groups}{{list of categories to group together}}
}
\value{
R data frame with posterior predictions for each category, across levels of xvar, all other variables held at mean.
}
\description{
Use category_groups to group categories together. The function asks the user to choose how to group
the categories together, in order to aid interpretation with many category models (e.g. ordinal or multinomial models).
}
