% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorMeans.R
\name{posterior_means}
\alias{posterior_means}
\title{Generate Posterior Predictions with 2 way interactions}
\usage{
posterior_means(
  model = regressionModels$pid3,
  xvar = "authM3",
  mvar = "latino",
  xval = seq(0, 1, 0.1),
  mval = c(0, 1)
)
}
\arguments{
\item{model}{{brms model}}

\item{xvar}{{independent variable}}

\item{mvar}{{moderator variable}}

\item{xrange}{{values of xvar}}

\item{mrange}{{values of mvar}}
}
\value{
R data frame with posterior predictions for each category, across levels of xvar, all other variables held at mean.
}
\description{
Use category_groups to group categories together. The function asks the user to choose how to group
the categories together, in order to aid interpretation with many category models (e.g. ordinal or multinomial models).
}
