% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorMarginal.R
\name{posterior_pme}
\alias{posterior_pme}
\title{Generate Predictive Marginal Effect}
\usage{
posterior_pme(
  model = regressionModels$envWater,
  xvar = "authM3",
  mvar = "latino",
  xrange = c(0, 1),
  mrange = c(0, 1)
)
}
\arguments{
\item{model}{{brms model}}

\item{xvar}{{independent variable}}

\item{mvar}{{moderator variable}}

\item{xrange}{{values of xvar}}

\item{mrange}{{values of mvar}}

\item{category_groups}{{list of categories to group together}}
}
\value{
R data frame with posterior predictions for each category, across levels of xvar, all other variables held at mean.
}
\description{
Use category_groups to group categories together. The function asks the user to choose how to group
the categories together, in order to aid interpretation with many category models (e.g. ordinal or multinomial models).
The function returns the average marginal effect for each category, across levels of xvar, all other variables observed.
}
