% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadDraws.R
\name{spreadDraw}
\alias{spreadDraw}
\title{Merge Draws from BRMS object to Secondary Data}
\usage{
spreadDraw(
  model1,
  mean_name = "mean_var",
  lower_name = "lower.DIF",
  upper_name = "upper.DIF"
)
}
\description{
Merge Draws from BRMS object to Secondary Data
}
