% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroOne.R
\name{zero.one}
\alias{zero.one}
\title{Create a zero one normalization}
\usage{
zero.one(x)
}
\arguments{
\item{numeric}{value}
}
\value{
A standardized 0-1 variable
}
\description{
Create a zero one normalization
}
\examples{
#rnorm(10) \%>\% zero.one()
}
